requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
faiss-cpu
sentence-transformers
numpy
pandas
pydantic
alembic